Make DashBoardOverview a separate component
