+++ frontend/main-app/src/components/Dashboard/Dashboard.module.css
@@ -1,62 +1,54 @@
 .dashboard {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: var(--fds-spacing-2);
-}
-
-.overview {
-  background-color: var(--fds-semantic-background-default);
-  border: 1px solid var(--fds-semantic-border-neutral-subtle);
-  border-radius: var(--fds-border_radius-medium);
-  padding: var(--fds-spacing-5);
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    gap: var(--fds-spacing-2);
 }
 
 .formList {
-  display: flex;
-  flex-direction: column;
-  gap: var(--fds-spacing-2);
+    display: flex;
+    flex-direction: column;
+    gap: var(--fds-spacing-2);
 }
 
-.accordionContentContainer,
-.headingContainer {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: var(--fds-spacing-4);
+.accordionContentContainer {
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: var(--fds-spacing-4);
 }
 
 .buttonContainer {
-  padding-top: var(--fds-spacing-1);
-  display: flex;
-  flex-direction: column;
-  gap: var(--fds-spacing-4);
+    padding-top: var(--fds-spacing-1);
+    display: flex;
+    flex-direction: column;
+    gap: var(--fds-spacing-4);
 }
 
 .button {
-  justify-content: left;
+    justify-content: left;
 }
 
 @media screen and (max-width: 1000px) {
-  .accordionContentContainer {
-    flex-direction: column;
-  }
+    .accordionContentContainer {
+        flex-direction: column;
+    }
 
-  .infoContainer {
-    margin-bottom: var(--fds-spacing-4);
-  }
+    .infoContainer {
+        margin-bottom: var(--fds-spacing-4);
+    }
 
-  .buttonContainer {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-  }
+    .buttonContainer {
+        display: grid;
+        grid-template-columns: 1fr 1fr;
+    }
 
-  .button {
-    justify-content: center;
-  }
+    .button {
+        justify-content: center;
+    }
 }
 
 @media screen and (max-width: 550px) {
-  .buttonContainer {
-    grid-template-columns: 1fr;
-  }
+    .buttonContainer {
+        grid-template-columns: 1fr;
+    }
 }
